more on eval

+++ evaluation/step0_extract_feature.sh
@@ -1,6 +1,6 @@
 #!/bin/bash -e
 # $File: step0_extract_feature.sh
-# $Date: Sun May 17 11:44:13 2015 +0800
+# $Date: Sun May 31 20:19:08 2015 +0800
 # $Author: jiakai <jia.kai66@gmail.com>
 
 FILE_LIST=../../sliver07/list-all.txt
@@ -8,6 +8,7 @@ OUTPUT_DIR=data/feature
 
 model_runner=$1
 model_name=$2
+shift 2 || true
 
 if [ ! -f "$FILE_LIST" ]
 then
@@ -17,19 +18,29 @@ fi
 
 if [ -z "$model_name" ]
 then
-    echo "usage: $0 <model runner> <model name>"
+    echo "usage: $0 <model runner> <model name> [-c]"
     exit -1
 fi
 
-
 output_dir=$OUTPUT_DIR/$model_name
-rm -rf $output_dir
-mkdir -pv $output_dir
-(echo "cmdline: $@"; date; env) > $output_dir/runtime_env
+
+[ "$1" == "-c" ] || rm -rf $output_dir
+
+if [ ! -d "$output_dir" ]
+then
+    mkdir -pv $output_dir
+    (echo "cmdline: $@"; date; env) > $output_dir/runtime_env
+fi
 
 for i in $(cat $FILE_LIST)
 do
     echo $i
-    $model_runner $(dirname $FILE_LIST)/$i \
-        $output_dir/$(basename $i | cut -d. -f1).pkl
+    cur_out=$output_dir/$(basename $i | cut -d. -f1).pkl
+    if [ -f "$cur_out" ]
+    then
+        echo "$cur_out already exists, ignore"
+        continue
+    fi
+    $model_runner $(dirname $FILE_LIST)/$i $cur_out
+
 done
